Fix stack overflow bug
Fix bug that caused stack overflow when parsing many consecutive line or
block comments.

The problem was that parse_whitespace() calls parse_comment() which
indirectly calls parse_whitepace() again. So each consecutive line or
block comment would add more frames to the stack. This makes it possible
for a malicious actor to easily craft a JSONC document that will cause
your program to exit with a segfault.